Add Authentication

+++ InventoryApp.API/Controllers/CategoryController.cs
@@ -1,7 +1,7 @@
-﻿using FluentValidation;
+﻿using System.Security.Claims;
+using FluentValidation;
 using InventoryApp.Service.DTO;
 using InventoryApp.Service.Interface;
-using InventoryApp.Service.Validators;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -23,16 +23,28 @@ namespace InventoryApp.API.Controllers
         [HttpGet("GetAll")]
         public async Task<IActionResult> GetAll([FromQuery] int? categoryId)
         {
+            if (GetUserIdFromToken() != 1)
+            {
+                return Unauthorized("You are not authorized to delete this product");
+            }
             return Ok(await _service.GetAll(categoryId));
         }
         [HttpGet("GetById")]
         public async Task<IActionResult> GetById([FromQuery] int id)
         {
+            if (GetUserIdFromToken() != 1)
+            {
+                return Unauthorized("You are not authorized to delete this product");
+            }
             return Ok(await _service.GetById(id));
         }
         [HttpPost("Create")]
         public async Task<IActionResult> Create([FromBody] CategoryDto dto)
         {
+            if (GetUserIdFromToken() != 1)
+            {
+                return Unauthorized("You are not authorized to delete this product");
+            }
             var validationResult = await _validator.ValidateAsync(dto);
 
             if (!validationResult.IsValid)
@@ -50,6 +62,10 @@ namespace InventoryApp.API.Controllers
         [HttpPut("Update")]
         public async Task<IActionResult> Update([FromQuery] int id, [FromBody] CategoryDto dto)
         {
+            if (GetUserIdFromToken() != 1)
+            {
+                return Unauthorized("You are not authorized to delete this product");
+            }
             var validationResult = await _validator.ValidateAsync(dto);
 
             if (!validationResult.IsValid)
@@ -71,12 +87,38 @@ namespace InventoryApp.API.Controllers
         [HttpPut("Delete")]
         public async Task<IActionResult> Delete([FromQuery] int id)
         {
+            if (GetUserIdFromToken() != 1)
+            {
+                return Unauthorized("You are not authorized to delete this product");
+            }
             var result = await _service.Delete(id);
             if (!result)
             {
                 return NotFound(new { message = "Category not found." });
             }
             return Ok("Category deleted successfully");
+        }
+        protected long GetUserIdFromToken()
+        {
+            long UserId = 0;
+            try
+            {
+                if (HttpContext.User.Identity.IsAuthenticated)
+                {
+                    var identity = HttpContext.User.Identity as ClaimsIdentity;
+                    if (identity != null)
+                    {
+                        IEnumerable<Claim> claims = identity.Claims;
+                        string strUserId = identity.FindFirst("UserId").Value;
+                        long.TryParse(strUserId, out UserId);
+                    }
+                }
+                return UserId;
+            }
+            catch
+            {
+                return UserId;
+            }
         }
     }
 }
